requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
DateTime==4.3
idna==2.9
pprint==0.1
praw==7.0.0
prawcore==1.4.0
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
six==1.15.0
spotipy==2.12.0
update-checker==0.17
urllib3==1.25.9
websocket-client==0.57.0
zope.interface==5.1.0
beautifulsoup4
pprint
pickle